+++ src/modules/algos/AlgosTypes.ts
@@ -4,6 +4,22 @@ import {
     PermissionsType,
 } from '@/modules/common/CommonTypes';
 
+export enum AssetKindT {
+    dataSample = 'ASSET_DATA_SAMPLE',
+    model = 'ASSET_MODEL',
+    dataManager = 'ASSET_DATA_MANAGER',
+}
+export type InputT = {
+    kind: AssetKindT;
+    multiple: boolean;
+    optional: boolean;
+};
+
+type OutputT = {
+    kind: AssetKindT;
+    multiple: boolean;
+};
+
 export interface AlgoT {
     key: string;
     name: string;
@@ -13,4 +29,6 @@ export interface AlgoT {
     algorithm: FileT;
     metadata: MetadataT;
     creation_date: string;
+    inputs: { [name: string]: InputT };
+    outputs: { [name: string]: OutputT };
 }